# perf_24x7_all_events - add support for POWER10 processor family