Running perf_24x7_all_events test on a POWER10 system fails with following 
error:
ERROR: 'hv_24x7_all_events' object has no attribute 'perf_stat' (0.84 s)

This test lacks support for POWER10 processor family. The patch adds that
support. It also fixes warnings reported by pycodestyle.

Signed-off-by: Sachin Sant <sachinp@linux.vnet.ibm.com>