@sacsant 2021-05-04
Without the fix:
 (1/1) perf/perf_24x7_all_events.py:hv_24x7_all_events.test_all_events: ERROR: 'hv_24x7_all_events' object has no attribute 'perf_stat' (0.84 s)
RESULTS    : PASS 0 | ERROR 1 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB TIME   : 1.06 s

With the fix (with --dry-run):
 (1/1) perf/perf_24x7_all_events.py:hv_24x7_all_events.test_all_events: CANCEL: Test cancelled due to --dry-run (0.00 s)
RESULTS    : PASS 0 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 1
JOB TIME   : 0.21 s