@nasastry 2021-05-06
> @nasastry Can you please review this?

@sacsant Code looks good, it would be better to send two patches one for power10 change and other for the code style change. 
Thanks!!