@sacsant 2021-05-06
> @sacsant Code looks good, it would be better to send two patches one for power10 change and other for the code style change.
Sure I can split the patch into two, one for P10 support and another for code style warning.